move chain under src with other processes + update docs with more details
