--- chain/small_geojson.json
@@ -1,36 +1,0 @@
-{
-  "type": "FeatureCollection",
-  "features": [
-    {
-      "type": "Feature",
-      "properties": {},
-      "geometry": {
-        "type": "Polygon",
-        "coordinates": [
-          [
-            [
-              -73,
-              45
-            ],
-            [
-              -72,
-              45
-            ],
-            [
-              -72,
-              46
-            ],
-            [
-              -73,
-              46
-            ],
-            [
-              -73,
-              45
-            ]
-          ]
-        ]
-      }
-    }
-  ]
-}
